+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Changelog
 * Add :meth:`fuc.api.pyvcf.VcfFrame.plot_titv` method.
 * Add ``palette`` and ``flip`` arguments to :meth:`fuc.api.pymaf.MafFrame.plot_vartype` method.
 * Update :meth:`fuc.api.pymaf.MafFrame.plot_snvclsc` method to support creation of a grouped bar plot.
+* Update :meth:`fuc.api.pymaf.MafFrame.plot_snvclsp` method to support creation of a grouped box plot.
 
 0.15.0 (2021-06-24)
 -------------------